+++ src/libs/index.ts
@@ -1,5 +1,4 @@
 export * from '@/libs/calculateTotalArticles'
-export * from '@/libs/cookies'
 export * from '@/libs/errors'
 export * from '@/libs/expression'
 export * from '@/libs/fetcher'
